requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
fastapi-x402>=0.1.0
anthropic>=0.40.0
httpx>=0.28.0
python-dotenv>=1.0.0
cachetools>=4.2.0
playwright>=1.45.0
Pillow>=10.4.0
